# api/openapi.v1.json
{
...
  "info": {
    "description": "Versioned HTTP contract for ForgeFlow integrations. Internal runtime state remains controller-owned.",
    "title": "ForgeFlow Control Plane API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/improvements": {
      "get": {
        "operationId": "improvementsList",
        "parameters": [
          {
            "in": "query",
            "name": "programId",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "DISCOVERED",
                "QUEUED",
                "ADOPTED",
                "REJECTED",
                "STALE",
                "COMPLETED"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 1000,
                  "minimum": 1,
                  "type": "integer"
                },
                {
                  "type": "string"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "count": {
                      "minimum": 0,
                      "type": "integer"
                    },
                    "items": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "candidateId": {
                            "type": "string"
                          },
                          "evidence": {
                            "items": {
                              "type": "string"
                            },
                            "type": "array"
                          },
                          "fingerprint": {
                            "type": "string"
                          },
                          "planId": {
                            "type": "string"
                          },
                          "programId": {
                            "type": "string"
                          },
                          "pullRequestId": {
                            "type": "string"
                          },
                          "risk": {
                            "enum": [
                              "LOW",
                              "MEDIUM",
                              "HIGH"
                            ],
                            "type": "string"
                          },
                          "status": {
                            "enum": [
                              "DISCOVERED",
                              "QUEUED",
                              "ADOPTED",
                              "REJECTED",
                              "STALE",
                              "COMPLETED"
                            ],
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "candidateId",
                          "programId",
                          "fingerprint",
                          "title",
                          "evidence",
                          "risk",
                          "status"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "runtime": {
                      "additionalProperties": true,
                      "type": "object"
                    }
                  },
                  "required": [
                    "items",
                    "count",
                    "runtime"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/improvements/{candidateId}": {
      "get": {
        "operationId": "improvementsGet",
        "parameters": [
          {
            "in": "path",
            "name": "candidateId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "candidate": {
                      "additionalProperties": false,
                      "properties": {
                        "candidateId": {
                          "type": "string"
                        },
                        "evidence": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "fingerprint": {
                          "type": "string"
                        },
                        "planId": {
                          "type": "string"
                        },
                        "programId": {
                          "type": "string"
                        },
                        "pullRequestId": {
                          "type": "string"
                        },
                        "risk": {
                          "enum": [
                            "LOW",
                            "MEDIUM",
                            "HIGH"
                          ],
                          "type": "string"
                        },
                        "status": {
                          "enum": [
                            "DISCOVERED",
                            "QUEUED",
                            "ADOPTED",
                            "REJECTED",
                            "STALE",
                            "COMPLETED"
                          ],
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "candidateId",
                        "programId",
                        "fingerprint",
                        "title",
                        "evidence",
                        "risk",
                        "status"
                      ],
                      "type": "object"
                    },
                    "diagnoses": {
                      "items": {
                        "additionalProperties": true,
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "plan": {
                      "anyOf": [
                        {
                          "additionalProperties": true,
                          "type": "object"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "program": {
                      "additionalProperties": false,
                      "properties": {
                        "autoMerge": {
                          "type": "boolean"
                        },
                        "autonomousScope": {
                          "enum": [
                            "CONSERVATIVE",
                            "STANDARD"
                          ],
                          "type": "string"
                        },
                        "candidateRisk": {
                          "enum": [
                            "LOW",
                            "MEDIUM",
                            "HIGH"
                          ],
                          "type": "string"
                        },
                        "enabled": {
                          "type": "boolean"
                        },
                        "failureCodePrefixes": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "failureThreshold": {
                          "type": "integer"
                        },
                        "implementationRoutes": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "programId": {
                          "type": "string"
                        },
                        "projectKey": {
                          "type": "string"
                        },
                        "recentExecutionLimit": {
                          "type": "integer"
                        },
                        "repositoryPath": {
                          "type": "string"
                        },
                        "reviewRoutes": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        }
                      },
                      "required": [
                        "programId",
                        "projectKey",
                        "autonomousScope",
                        "autoMerge",
                        "enabled"
                      ],
                      "type": "object"
                    },
                    "selfChange": {
                      "additionalProperties": false,
                      "properties": {
                        "canaries": {
                          "items": {
                            "additionalProperties": true,
                            "type": "object"
                          },
                          "type": "array"
                        },
                        "promotion": {
                          "anyOf": [
                            {
                              "additionalProperties": true,
                              "type": "object"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "promotionRequest": {
                          "anyOf": [
                            {
                              "additionalProperties": true,
                              "type": "object"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "releaseProvenance": {
                          "anyOf": [
                            {
                              "additionalProperties": true,
                              "type": "object"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "selfChange": {
                          "type": "boolean"
                        },
                        "selfPromotionEnabled": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "selfChange",
                        "selfPromotionEnabled",
                        "canaries",
                        "promotionRequest",
                        "promotion",
                        "releaseProvenance"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "candidate",
                    "program",
                    "plan",
                    "diagnoses",
                    "selfChange"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/improvements/{candidateId}/adopt": {
      "post": {
        "operationId": "improvementsAdopt",
        "parameters": [
          {
            "in": "path",
            "name": "candidateId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "additionalProperties": true,
                    "properties": {
                      "acknowledgeHighRisk": {
                        "type": "boolean"
                      },
                      "baseRevision": {
                        "type": "string"
                      },
                      "delivery": {
                        "additionalProperties": true,
                        "properties": {
                          "autoMerge": {
                            "type": "boolean"
                          },
                          "branch": {
                            "type": "string"
                          },
                          "mergeMethod": {
                            "enum": [
                              "merge",
                              "squash",
                              "rebase"
                            ],
                            "type": "string"
                          },
                          "remote": {
                            "type": "string"
                          },
                          "requiredChecks": {
                            "items": {
                              "type": "string"
                            },
                            "type": "array"
                          },
                          "targetBranch": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "branch",
                          "autoMerge"
                        ],
                        "type": "object"
                      },
                      "priority": {
                        "type": "integer"
                      },
                      "repositoryPath": {
                        "type": "string"
                      }
                    },
                    "type": "object"
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            }
          },
          "required": false
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "properties": {
                    "candidate": {
                      "additionalProperties": false,
                      "properties": {
                        "candidateId": {
                          "type": "string"
                        },
                        "evidence": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "fingerprint": {
                          "type": "string"
                        },
                        "planId": {
                          "type": "string"
                        },
                        "programId": {
                          "type": "string"
                        },
                        "pullRequestId": {
                          "type": "string"
                        },
                        "risk": {
                          "enum": [
                            "LOW",
                            "MEDIUM",
                            "HIGH"
                          ],
                          "type": "string"
                        },
                        "status": {
                          "enum": [
                            "DISCOVERED",
                            "QUEUED",
                            "ADOPTED",
                            "REJECTED",
                            "STALE",
                            "COMPLETED"
                          ],
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "candidateId",
                        "programId",
                        "fingerprint",
                        "title",
                        "evidence",
                        "risk",
                        "status"
                      ],
                      "type": "object"
                    },
                    "plan": {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    "scheduling": {
                      "additionalProperties": true,
                      "type": "object"
                    }
                  },
                  "required": [
                    "candidate",
                    "plan"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/improvements/{candidateId}/diagnose": {
      "post": {
        "operationId": "improvementsDiagnose",
        "parameters": [
          {
            "in": "path",
            "name": "candidateId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "diagnosis": {
                      "additionalProperties": true,
                      "type": "object"
                    }
                  },
                  "required": [
                    "diagnosis"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/improvements/{candidateId}/reconcile": {
      "post": {
        "operationId": "improvementsReconcile",
        "parameters": [
          {
            "in": "path",
            "name": "candidateId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "candidate": {
                      "additionalProperties": false,
                      "properties": {
                        "candidateId": {
                          "type": "string"
                        },
                        "evidence": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "fingerprint": {
                          "type": "string"
                        },
                        "planId": {
                          "type": "string"
                        },
                        "programId": {
                          "type": "string"
                        },
                        "pullRequestId": {
                          "type": "string"
                        },
                        "risk": {
                          "enum": [
                            "LOW",
                            "MEDIUM",
                            "HIGH"
                          ],
                          "type": "string"
                        },
                        "status": {
                          "enum": [
                            "DISCOVERED",
                            "QUEUED",
                            "ADOPTED",
                            "REJECTED",
                            "STALE",
                            "COMPLETED"
                          ],
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "candidateId",
                        "programId",
                        "fingerprint",
                        "title",
                        "evidence",
                        "risk",
                        "status"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "candidate"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/improvements/{candidateId}/reject": {
      "post": {
        "operationId": "improvementsReject",
        "parameters": [
          {
            "in": "path",
            "name": "candidateId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "candidate": {
                      "additionalProperties": false,
                      "properties": {
                        "candidateId": {
                          "type": "string"
                        },
                        "evidence": {
                          "items": {
                            "type": "string"
                          },
                          "type": "array"
                        },
                        "fingerprint": {
                          "type": "string"
                        },
                        "planId": {
                          "type": "string"
                        },
                        "programId": {
                          "type": "string"
                        },
                        "pullRequestId": {
                          "type": "string"
                        },
                        "risk": {
                          "enum": [
                            "LOW",
                            "MEDIUM",
                            "HIGH"
                          ],
                          "type": "string"
                        },
                        "status": {
                          "enum": [
                            "DISCOVERED",
                            "QUEUED",
                            "ADOPTED",
                            "REJECTED",
                            "STALE",
                            "COMPLETED"
                          ],
                          "type": "string"
                        },
                        "title": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "candidateId",
                        "programId",
                        "fingerprint",
                        "title",
                        "evidence",
                        "risk",
                        "status"
                      ],
                      "type": "object"
                    }
                  },
                  "required": [
                    "candidate"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          }
        }
      }
    },
...
  },
...
}